python/requirements.txt
python-dotenv
Pillow
google-generativeai
mistralai
httpx
appdirs
opencv-python
jinja2
pyinstaller
imageio-ffmpeg
